+++ package/Runtime/Hotspots/Scripts/Needle.Hotspots.codegen/HotspotBehaviour.cs
@@ -1,0 +1,27 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class HotspotBehaviour : UnityEngine.MonoBehaviour
+	{
+		public UnityEngine.UI.Text @label;
+		public UnityEngine.UI.Text @content;
+		public UnityEngine.GameObject @shift;
+		public float @zOffset;
+		public float @contentFadeDuration;
+		public float @hotspotFadeDuration;
+		public UnityEngine.Canvas @canvas;
+		public UnityEngine.CanvasGroup @contentCanvasGroup;
+		public UnityEngine.CanvasGroup @headerCanvasGroup;
+		public void init(Needle.Typescript.GeneratedComponents.Hotspot @hotspot){}
+		public void onPointerClick(){}
+		public void deselect(){}
+		public void onBeforeRender(object @frame){}
+		public void applyFade(){}
+	}
+}
+
+// NEEDLE_CODEGEN_END