+++ package/Runtime/Scrollytelling/Needle.Samples.Scrollytelling.codegen/ScrollTimeline_2.cs
@@ -1,0 +1,18 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class ScrollTimeline_2 : UnityEngine.MonoBehaviour
+	{
+		public UnityEngine.Playables.PlayableDirector @timeline;
+		public float @scrollSpeed = 0.5f;
+		public float @lerpSpeed = 2.5f;
+		public void start(){}
+		public void onBeforeRender(){}
+	}
+}
+
+// NEEDLE_CODEGEN_END