+++ package/Runtime/Sidescroller/Needle.Sidescroller.codegen/SidescrollerInfo.cs
@@ -1,0 +1,15 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class SidescrollerInfo : UnityEngine.MonoBehaviour
+	{
+		public void OnEnable(){}
+		public void OnDisable(){}
+	}
+}
+
+// NEEDLE_CODEGEN_END