+++ package/Runtime/_Shared/SampleScripts.codegen/DisableEnvironmentLight.cs
@@ -1,0 +1,15 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class DisableEnvironmentLight : UnityEngine.MonoBehaviour
+	{
+		public void OnEnable(){}
+		public void OnDisable(){}
+	}
+}
+
+// NEEDLE_CODEGEN_END