+++ package/Runtime/_Shared/SampleScripts.codegen/PrefabSceneSwitcherSample.cs
@@ -1,0 +1,14 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class PrefabSceneSwitcherSample : UnityEngine.MonoBehaviour
+	{
+		public UnityEngine.Transform[] @scenes;
+	}
+}
+
+// NEEDLE_CODEGEN_END