+++ package/Changelog.md
@@ -4,8 +4,19 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.5.0] - 2023-03-03
+## [0.6.0] - 2023-04-11
+- add PostProcessing sample
+- add Sidescroller sample
+- add Vuejs sample
+- add Hotspots sample
+- add Multi-Scene sample
 
+## [0.5.1] - 2023-03-14
+- add new GroundProjection sample
+- fix script import paths for `2.65.2`
+- fix iframe click not being received sometimes
+
+## [0.5.0] - 2023-03-03
 - fixed some sample materials not properly updating in 2021.3 and 2022.1
 - fixed screensharing sample having a broken Prefab reference
 - fixed Web Projects missing for HTML sample and Scrollytelling sample
@@ -13,5 +24,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - added particle collision sample
 
 ## [0.4.0] - 2022-12-16
-
 - initial registry release